requirements.txt
aniso8601==10.0.1
blinker==1.9.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
flake8==7.2.0
Flask==2.2.5
flask-mongoengine==1.0.0
Flask-RESTful==0.3.10
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mongoengine==0.29.1
pycodestyle==2.13.0
pyflakes==3.3.2
pymongo==4.13.0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
Werkzeug==2.3.8
WTForms==3.2.2
pytest==8.3.5